chore: lint config package (#3382)
The main goal of this change is to add linting for the `config` package.
While doing this I wanted to take the chance and add a few more things:

1. Root `Makefile` for root operations `install` and `clean` for easy
cleaning and reinstallation of dependencies
2. Add rule `import/no-extraneous-dependencies` to error in case we make
use of a module without installing it beforehand. This caused some
issues already and a first fix was added with #3381. But also here I
have to install `glob` as it wasn't listed in the config's dependency
list, but is imported in
[packages/config/scripts/ci.cjs#L1](https://github.com/kumahq/kuma-gui/blob/36911e6e113bdf823e98233d2eac4543acb588b4/packages/config/scripts/ci.cjs#L1).
3. Linting of staged files in `packages/config`
4. Auto-fixing some linting issues that got revealed after adding the
linter here
5. A typo I found

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>