chore(deps): upgrade reflect-metadata (#3384)
FYI: You will only see a warning for the issue here if you run `make
install` with completely clean `node_modules`

---

I chose to put the override in the root package.json, I also had to run
`npm dedupe` several times to get it to "stick" so unfortunately quite a
lot of churn in the package-lock.json.

I just looked at it myself, and there are lots of downgrades of some
transitive dependencies here (as well as upgrades). I would assume these
still satisfy the semver ranges for everything and are a consequence of
the dedupe's.

Ideally I'd like less churn here though, so if there are any
suggestions/ideas for trying to get a smaller changeset I'd be very
grateful. Otherwise I think this is probably still good as it seems like
the changed deps are mostly (at least) dev/tooling deps.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>