fix: flaky tests due to mismatch of payload type (#3419)
Over the past weeks we have noticed some flakes of our browser tests
where the assertion was mainly about a state being `online`. I have
traced it down to a mismatch of our payload type. In our tests we always
set the state that we want to assert for in
`dataplane.inbound.health.state`, but according to our type in
[data-planes](https://github.com/kumahq/kuma-gui/blob/26976933fb563768d5f09099f2d7d28c11545bcd/packages/kuma-gui/src/app/data-planes/data/DataplaneNetworking.ts#L13)
there is no entry `health`. The `state` can be found in
`dataplane.inbound.state`.

Example:
![Screenshot 2025-01-20 at 11 11
05](https://github.com/user-attachments/assets/0df626fb-d6d2-4efa-889d-fe0a84c8a83f)

Closes #3376

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>