refactor: for route state prefer events over v-model (#3416)
In #1966 a watcher has been added to sync boolean query search params.
This watcher caused some issues in nested routes due to a race condition
as described in #3409.

The reason for the watcher (from #1966)
>Whilst I'd prefer not use two way binding models we have some
components that we use that uses these as required values, such as
KInputSwitch

Turns out that the components `KSelect` and `KCheckBox` now support the
`@change` events. When used alongside the `:model-value` we have our
getter and setter separated. This means that we can remove the
additional watcher.

From #3409 
> 4. I had to keep the watcher with the booleans only for the case of
KInputSwitch.

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>