fix: changes Xi18n's default-message prop to default-path (#3488)
We noticed this:

![image
(3)](https://github.com/user-attachments/assets/7499366f-2cae-428b-9050-d69f3242aa7b)

which is the thing we are fixing here, but its a bit of a longer story
due to what Xi18n should be doing.

---

The bug is that we aren't using a `:` here:


https://github.com/kumahq/kuma-gui/blob/603c39571c34fcdfad9eeafccf02ee4453fdc17b/packages/kuma-gui/src/app/policies/views/PolicyListView.vue#L68

At first I was like "I can just add the `:` and job done", but then I
was like "wait a second that means I can pass unencoded text into
`default-message` using a normal `t` and `XI18n` will blindly render it,
therefore bypassing any encoding.

This means we can't give `XI18n` a `default-message` prop, because we
can't guarantee that its encoded correctly (i.e. sometimes we can use
`t` which will encode it, but if we need some HTML in the string that
`t` returns, then we can't guarantee and params passed to that are
encoded correcly.

I chose to make `default-message` into `default-path` (which matches the
`path` prop), i.e. you don't pass it a message to use for a default, you
pass it a path, which is then looked up internally.

This means two things:

- You can't currently pass parameters to a default path (we don't need
to, yet at least). We don't use a default very much so its no big deal
later down the line to add a `default-path={ path 'common.something',
params: {}}` or just an additional `default-params` props, or something.
- We need a i18n string that is literally a blank string i.e.
`common.i18n.ignore-errors`. ... I'm actually wondering now if I should
change this PR to make `ignoreErrors` an option instead 🤔 of the
`common.i18n.ignore-errors` thing.

@schogges what do you think on this last bit, the more I think about it
we should just make an `ignoreError` option or `ignoreMissingPath` or
something, wdyt? Lmk if you don't follow the thinking here

Signed-off-by: John Cowen <john.cowen@konghq.com>