fix: removes extra padding adding via kongponents update (#3493)
Several people have reported issue with being able to click on our tab
component i.e. you have to click exactly on the text instead of being
able to click on the entire button.

A recent kongponent update
(https://github.com/Kong/kongponents/pull/2532) changed the padding of
certain HTML elements, which removed the padding from our anchor links
and added it to the surrounding div instead, meaning whilst it looked
like you could click on the entire button to click the anchor, you where
actually clicking the div, meaning the anchor wasn't navigating.

Also see
https://github.com/Kong/kongponents/pull/2532/files#diff-01f83cc2092a263bfd47c1d0a4ba5e96dea98264d85282662a63f4fa18c44dedR177

This PR rearranges the padding using some overwrites to return the
anchor its previous padded state.

There is still a little unclickable area between the button and the
bottom border, but I think this is intentional form the kongponents
authors. The additional difference with this area is that at least you
don't see a pointer cursor when hovering over it.

Signed-off-by: John Cowen <john.cowen@konghq.com>