feat(data-plane-inbounds): show port name (#3483)
This adds the optionally existing name of the inbound port. Currently
there is a conflict with an already existing `name` property on the
inbounds, therefore I went for `portName`. But I've also added it to the
summary including the ability to copy.

---

For now I left out the ability to copy the name because there is an
issue with the tooltip popover. It's based on floating-ui and the
position is not properly calculated when the trigger is nested in an
element with `container-type: inline-size`.

![Screenshot 2025-01-28 at 15 06
58](https://github.com/user-attachments/assets/06a18ba8-c192-4b20-aa0d-f4222fc7d806)


![image](https://github.com/user-attachments/assets/3e8bd465-97f6-4d1e-80a1-363d3d734f8d)

Closes #3451

Signed-off-by: schogges <moritz.fleck@konghq.com>