feat(vcp-health): add policies, show totals only, hide border on wrapped items (#3490)
Transforming the VCP health card into a VCP status card, that show total
numbers. Additionally I've aligned the icons with the navigation and
added the total number of policies to the group.


![image](https://github.com/user-attachments/assets/5b993df4-ba2c-452e-8309-c77d6ffae40e)

---

As part of this I've also fixed an issue where the borders
(`.columns-with-borders`) would still be shown when the items wrap and
the layout turns from row into column layout. Now when the container is
smaller than `640px`, the borders will be hidden.

Closes #2667

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>