chore(dataplanes): change all refs of route.dataPlane > route.proxy (#3512)
Ideally we want to reuse some views used in dataplanes with some views
used in zone-ingress/egrees (i.e. ZoneProxies).

If we stand any chance of doing this without making some serious changes
elsewhere, we need to make some more things similar.

This PR changes all instances of the `route.params.dataPlane` route
parameter to instead use `route.params.proxy`. Back in
https://github.com/kumahq/kuma-gui/pull/3509 we changed all the "zone
proxy" views to also use the same name. At the end of the day Zone
Proxies are just dataplanes i.e. they are essentially the same thing (an
instance of Envoy).

I'm honestly not totally sure to what extent we'll be able to reuse
still (edit: I found doing this is very helpful 😸 ), but at the very
least this change gets rid of an awkward casing of `dataPlane`. `proxy`
is better as its a single word and there is no chance of getting weird
casing differences like `pRoXy`

Signed-off-by: John Cowen <john.cowen@konghq.com>