feat: wrap nested route features in x-card (#3518)
In addition to wrap tables nested in routes in `XCard`, I decided to
wrap all nested features in `XCard`. Also some of the titles of these
features where not part of the `XCard` and I moved them to the `#title`
slot. This contributes to overall consistency.

---

Apart from moving things to `XCard` I moved the warning about missing
tls in the `DataPlaneDetailView` to the other warnings, which improves
visibility on that as it's now at the top of the page.


![image](https://github.com/user-attachments/assets/5d07cba2-1030-412d-b692-c26dc9e80785)

Some more examples:


![image](https://github.com/user-attachments/assets/71ed07db-7a3b-4e34-9e74-a291c84c4146)

![image](https://github.com/user-attachments/assets/03a402cd-2be6-4e56-8f87-326f61aab77e)

Closes #3453

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>