feat(cp-stats): add mesh-services (#3551)
This introduces `Mesh*Services` to our vcp stats. As it can happen that
we have a mix of `Mesh*Services` and legacy services in the meshes, I
added the possibility to show both in the same `ResourceStatus`
including a `#description`.
To achieve this I've made the `#title`-slot optional and added an
optional `#description`-slot. With this it's possible to nest multiple
`ResourceStatus` in a "main" `ResourceStatus` which keeps the same
layout and styling and also the declarative approach of rendering
things.

In case there are no `Mesh*Services` it only renders the number of the
legacy services. In case there are no legacy services, but
`Mesh*Services` it renders the number of `Mesh*Services` only.
Only in case there is a mix of both it renders the `description` for
both including an info about the legacy services.

![Screenshot 2025-02-18 at 14 43
13](https://github.com/user-attachments/assets/cfebf188-ef50-4eaa-8dca-0ddf6ae97ca0)

![image](https://github.com/user-attachments/assets/c1f0463c-daf7-4f84-9446-26eb3153ab85)
![Screenshot 2025-02-17 at 12 31
38](https://github.com/user-attachments/assets/e5fe3160-cffb-4873-900b-edd0e8f5ca2a)

---

Regarding `contain: layout`:
This is required to prevent the issues regarding wrong positioning of
tooltips in an element of container-type: inline-size. Found this in
[this
article](https://dev.to/michaelcharles/chrome-129s-container-query-change-2i77).

Worth to mention that this solution is still not ideal. It creates a new
stacking context and may require to set additional z-index for elements
like tooltips that should overlay all other elements.


Closes #2966

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>