refactor(policies): use _resources instead of policies endpoint (#3568)
Instead of using the `/policies` endpoint we are now switching over to a
more general `/_resources` endpoint that returns a list of all
resources. The `/_resources` endpoint is typed through our openapi
specs, so I could remove the custom types.
In order to keep the changes as minimal as possible, I've used the
data-layer in the policies module to extract the policy types into a
separated `policyTypes` entry, which makes it easier to use in the
policy templates. Therefore I've also kept the source `/policy-types`
and didn't rename it according to the changes.
The `/_resources` endpoint could be reused in other parts of the app
that focuses on the non-policy entries and have a separate
implementation in the respective data-layer.

---

For the mocks I've extended the `resourceNames` with more resource names
and added an optional parameter to return all. As per default it returns
a random subset. To be able to differentiate between policies and
non-policies I've reused the kuma faker `policyName`, moved the items to
a `policyNames` function such that the policy names are in one place.

Closes #3565

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>
Signed-off-by: Mo <24444046+schogges@users.noreply.github.com>
Co-authored-by: John Cowen <johncowen@users.noreply.github.com>