feat(policy-summary): add spec to structured view (#3579)
This is the first part and in a follow-up I'll add more view types for
the config. Here I added the spec in a code block to the structured
view, so the user can still get all information in this view and doesn't
need to switch to YAML.

---

I added a prop to hide the `Copy as K8s` button from the code block. I
think the spec is agnostic to universal and k8s.

---


![image](https://github.com/user-attachments/assets/d7449a69-911f-405f-9be5-2c3e844a1fd0)

Part of #3489

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>