chore(lint): add CSS selector-combinator-space-after/before always lint (#3600)
Adds a CSS lint rule to force spaces around selector combinator.

Just to note, this is a code style rule, but at least we all agree on it
😅 .

It also adds a dev dependency 🙈 , so if that dependency ends up in our
weekly dependabot's all the time I am 100% ok with removing it again
(along with the rule).

Closes https://github.com/kumahq/kuma-gui/issues/1484

Signed-off-by: John Cowen <john.cowen@konghq.com>