refactor(meshes): slightly refactors MTP notification (#3599)
Refactors the MTP notification we have when you have mTLS enabled but no
MeshTrafficPermission.

---

While I was doing this I noticed that the faker `policyNames` code
wasn't quite the same interface as faker, and I think the randomness
wasn't quite working because max was defaulting to a super big value
(which I think meant a low chance of getting low numbers) so I made some
slight changes there after I looked in fakers code to see exactly how
they do this and then made a `maxmin` helper to replicate fakers
types/functionality so we keep the same interface everywhere.

This meant:

1. Altering slightly the `{ min, max }` args and how they are defaulted.
2. Using `Number.MAX_SAFE_INTEGER` over `Number.MAX_VALUE` as a "give me
all the things" placeholder
3. If you use `Number.MAX_SAFE_INTEGER` to say "potentially include all
of them" we don't pass that along to arrayElements
4. I removed includeLegacy where it wasn't needed, mostly because where
i was using it I had to write `(undefined, { includeLegacy: true})`

I _think_ this is the functionality we want for `policyNames`

---

Apart from this I also made the policy enabling/disabling rendering work
based on `total: 0` as well as just existence, which again is the
functionality I _think_ we want?

Closes https://github.com/kumahq/kuma-gui/issues/2851

Signed-off-by: John Cowen <john.cowen@konghq.com>