refactor(policy-summary): always show structured format option (#3625)
To improve on consistencies we decided to always offer the option
`structured` as a format in summary drawers. Currently that was missing
in `PolicySummaryView` and `DataPlanePolicySummaryView`. Now we just
want to only conditionally render the codeblock when there is no `spec`
available, which is only the case for legacy policies on universal.

Furthermore this adds the policy `type` to the structured view, such
that users can always see the type of the policy that they are looking
at.

Closes #3608

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>
Signed-off-by: John Cowen <john.cowen@konghq.com>
Co-authored-by: John Cowen <john.cowen@konghq.com>