chore(ci): iterative improvements to releasing (#3607)
Further iterates on our `release.yml` workflow. The intention here is to
simplify it enough so it literally calls `make build` and `make
release`. The only reason we split these in two is so we can use GH
Actions to generate an SBOM instead of the Makefile itself.

This approach means that we can potentially loop through several
packages in our workspace building and releasing each one with scripts
controlled by the Makefile itself, not CI YAML (which for reasons cannot
be totally versioned)

Separately I also added a similar Makefile target so you can test verify
the SBOM locally using `syft`. Note: This is not what we use in CI and
is only for testing/verification purposes.

Lastly, this would be good to get in the before we cut for release so
that moving forwards subsequent branches only need to support `make
build` and `make release`, but no big deal if this slips into 2.11

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>