fix(deps): pin entities-shared to 3.19.3 due to bad 3.20.0 (#3704)
We've noticed that various kong dependencies are being published with
the string `"workspace:^"` in their `peerDependencies`. This is causing
npm install to fail.

We've seen this specifically happening in
`@kong-ui-public/entities-shared@3.20.0` but it could also be happening
in other modules/version.

So at the very least we are pinning our version to not go past 3.19.3
until a fix is pushed upstream.

Signed-off-by: John Cowen <john.cowen@konghq.com>