refactor: move vite plugin, cypress interceptor and msw handlers to fake-api package  (#3693)
Creates a new package `@kumahq/fake-api` that serves a vite plugin, msw
handlers and cypress interceptor to be used as a server for our mocks.
This is still opt-in via `KUMA_API_URL=http://localhost:8080` and
`KUMA_MOCK_API_ENABLED=false`.

In order to load `.ts` files without compilation I found two options:

- From https://vite.dev/config/#configuring-vite there is
`--configLoader runner` which transforms given files on the fly. This
worked great for `vite` only actions, but failed when trying to run the
unit tests with `vitest`
- From https://nodejs.org/en/learn/typescript/run-natively using the
flag `--experimental-strip-types`, which works without any further
configuration on `vite`/`vitest`.

I decided to go for `--experimental-strip-types` only because it works
for both `vite` and `vitest`.

Part of #2001

---------

Signed-off-by: schogges <moritz.fleck@konghq.com>