fix(deps): don't use peerDependencies, we don't have consumers (#3708)
When we made `@kumahq/config` we started adding peerDependencies to
define the version of the "hosts" that we provide config/plugins for.

The thing is as dependbot (rightly) doesn't touch peerDependencies, and
the node ecosystem can sometimes move quite fast (major bumps are a
common occurrence), this means that our peerDependencies can quite
quickly get out of sync with what we have defined as
root/application/host dependencies. This is causing issues and
unnecessary work to keep these in sync manually (no dependbot).

Moreover, we are in total control of the `@kumahq/config` _and_ our host
app, and we don't publicly publish `@kumahq/config`, so in reality
whenever we bump the dependency in our host app (that needs to be in
sync with the peerDependencies) we should be bumping the
peerDependencies to the exact same version. This then begs the question,
why are we bumping in multiple places, when we can just bump in one.
i.e. in our unique situation what's the point of defining
peerDependencies if we are in full control of our entire codebase and we
don't provide a comsumable package to anyone else (where
peerDependencies would then become useful)

Theres probably more to say here, but the above is pretty much the gist
of it. All in all we decided we would remove and not use
peerDependencies at all.

This PR removes the peerDependencies we were using and moves what we had
there into the correct places.

Whilst doing this I spotted that we use certain cypress plugins in two
places just "because Cypress" so I also did the necessary work to make
it so we could specify these plugins in one single place in
`@kumahq/config` instead of also in `@kumahq/kuma-gui`.

As discussed over the past few days, there may end up being places where
we need to ensure a dependency is either in the root of the monorepo, or
in its individual workspace/application as a result of removing
peerDependencies. If/as and when we discover those we just move them as
and when.

Following this we said we would reassess in a couple of months to see
what effect this has had on our projects.

As a result of this PR the only cypress related dependencies you need to
install in one of our applications that needs e2e testing is `cypress`
itself and `@badeball/cypress-cucumber-preprocessor`, which we use in
our `steps` file. At some point our steps file will also be moved out of
the application and into a shareable packge/module. At which point all
an application will need to install locally is `cypress` itself plus
`@kumahq/config`

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>