fix(codeblock): fix height of codeblock to use DOM recycling scrolling (#3711)
Also see https://github.com/kumahq/kuma-gui/issues/3152, specifically
https://github.com/kumahq/kuma-gui/issues/3152#issuecomment-2758069136

---

Looks like the line numbers we render here can cause a massive slowdown
in re-renders, specifically when typing to search large configs.

The upstream KCodeBlock that we use for this does have a virtual
scroller for the line numbers built in, but because of the nature of
some virtual scrollers, and definitely in this case, you need to fix the
height of the container of the thing that contains the virtual scroller.
This means that the KCodeBlock now controls the scrolling, not the
browser window.

Note: I only implemented this in the "full page" XDS Config pages for
now, so if you are viewing the PR preview make sure you are looking at
it on this page.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>