chore(deps): move husky and lint-staged to root (#3687)
I hit a case where I installed the dependencies and `lint-staged` ended
up in the `packages/kuma-gui/node_modules` (not in the root). This
caused that the pre-commit hook failed: `.husky/pre-commit: line 1:
lint-staged: command not found`

We decided to move certain dependencies, that operate on the whole repo
or are meant to be added to all packages (i.e. `eslint`) to the
repository root as `devDepdendencies`. Therefore I am moving `husky` and
`lint-staged` to the repository root now.

Signed-off-by: schogges <moritz.fleck@konghq.com>