chore(notifications): normalize notification keys (#3728)
- Changes all out notifications to use lowercase i18n keys that all live
on a `<module-name>.notifications` property.
- Simplifies notifications so that are "calculated" in place instead of
some in the data layer and some in the view layer. They are now all in
the view layer. This means we can be more intentional about what
notifications are shown where. The is a small bug here where we try to
show a NACK warning on the zone listing page when we shouldn't.
- Due to the above point, I made some amends to the dataplane data layer
- Noticed a few tiny things in HostnameGenerators data mainly around
using the generated types in the mocks rather than our application
types.

Note tests are gonna fall on their face here, I wanted to put up a draft
PR to share progress, my next steps is fixing up the tests.

Part of https://github.com/kumahq/kuma-gui/issues/3605

I'm considering making a reusable/include-like component here, possibly
for the DataplaneListing, still considering it but it shouldn't
necessarily block review/merge, we can do a follow up.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>