chore: move XNotification to use :notify vs v-if (#3749)
Part of https://github.com/kumahq/kuma-gui/issues/3605

In the above issue we said:

> We need the dismissal actions to not only track when you have
dismissed a notification and store that, but also remove that storage
once the logic that showed the notification is reversed.

In order to know whether to show the notification we were using a `v-if`
to show choose whether to show the notification or not. But given the
above requirement we also need to remove the fact that you dismissed a
notification if the conditional logic is reversed. This means if we
continue to use `v-if` we can't do anything extra (like remove the
localstorage value) from within the XNotification if we don't even
render it.

This PR moves all notifications to use a `:notify` property, and moves
the v-if inside the component based to decide whether to render anything
based on this. As this boolean is now available to the component, if its
`false` and we have a localStorage value for this notification, we know
to remove it ready for the next time `:notify` is `true`.

Doing the localStorage manipulation will be a follow up PR concentrating
on the XNotification component itself rather than the callsites for the
XNotification.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>